adding h265_slice_header_t
(Has no meaning, sience we don't know the exact structure of the slice header. It is not publicly shared on the internet. You can find the ISO standard, but the documentation is not free. Btw the header would contain data for frame identification and slice (or frame part) identification, which would be necessary to solve our task.